requirements.txt
langchain
langchain-community
langchain-openai
openai
python-dotenv
streamlit
faiss-cpu
pypdf
tiktoken
